mercadolibre/requirements310.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.5
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
ez-setup==0.9
frozenlist==1.3.0
futures==3.0.5
futures3==1.0.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.3.7
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
zipp==3.8.0
